Getting rid of TestAstUtils redundancy
