+++ src/test/kotlin/cacophony/TestAstUtils.kt
@@ -1,6 +1,7 @@
 package cacophony
 
 import cacophony.semantic.CallGraph
+import cacophony.semantic.MAIN_FUNCTION_IDENTIFIER
 import cacophony.semantic.syntaxtree.*
 import cacophony.utils.Location
 
@@ -60,64 +61,52 @@ fun call(variableUse: VariableUse, vararg arguments: Expression) = FunctionCall(
 
 fun call(identifier: String, vararg arguments: Expression) = call(variableUse(identifier), *arguments)
 
-fun astOf(vararg expressions: Expression) = Block(mockRange(), expressions.toList())
+fun astOf(vararg expressions: Expression) =
+    Block(
+        mockRange(),
+        listOf(
+            functionDeclaration(
+                MAIN_FUNCTION_IDENTIFIER,
+                Block(mockRange(), expressions.toList()),
+            ),
+            call(variableUse(MAIN_FUNCTION_IDENTIFIER)),
+        ),
+    )
 
 fun callGraph(vararg calls: Pair<Definition.FunctionDeclaration, Definition.FunctionDeclaration>): CallGraph =
     calls.groupBy({ it.first }, { it.second }).mapValues { it.value.toSet() }
 
-fun addition(lhs: Expression, rhs: Expression) = lhs add rhs
-
 infix fun Expression.add(rhs: Expression) = OperatorBinary.Addition(mockRange(), this, rhs)
-
-fun subtraction(lhs: Expression, rhs: Expression) = lhs sub rhs
 
 infix fun Expression.sub(rhs: Expression) = OperatorBinary.Subtraction(mockRange(), this, rhs)
 
-fun multiplication(lhs: Expression, rhs: Expression) = lhs mul rhs
-
 infix fun Expression.mul(rhs: Expression) = OperatorBinary.Multiplication(mockRange(), this, rhs)
-
-fun division(lhs: Expression, rhs: Expression) = lhs div rhs
 
 infix fun Expression.div(rhs: Expression) = OperatorBinary.Division(mockRange(), this, rhs)
 
-fun modulo(lhs: Expression, rhs: Expression) = lhs mod rhs
-
 infix fun Expression.mod(rhs: Expression) = OperatorBinary.Modulo(mockRange(), this, rhs)
-
-fun minus(expression: Expression) = OperatorUnary.Minus(mockRange(), expression)
 
 infix fun Expression.land(rhs: Expression) = OperatorBinary.LogicalAnd(mockRange(), this, rhs)
 
 infix fun Expression.lor(rhs: Expression) = OperatorBinary.LogicalOr(mockRange(), this, rhs)
 
-fun equals(lhs: Expression, rhs: Expression) = lhs eq rhs
-
 infix fun Expression.eq(rhs: Expression) = OperatorBinary.Equals(mockRange(), this, rhs)
-
-fun notEquals(lhs: Expression, rhs: Expression) = lhs neq rhs
 
 infix fun Expression.neq(rhs: Expression) = OperatorBinary.NotEquals(mockRange(), this, rhs)
 
-fun lessEqual(lhs: Expression, rhs: Expression) = lhs leq rhs
-
 infix fun Expression.leq(rhs: Expression) = OperatorBinary.LessEqual(mockRange(), this, rhs)
-
-fun lessThan(lhs: Expression, rhs: Expression) = lhs lt rhs
 
 infix fun Expression.lt(rhs: Expression) = OperatorBinary.Less(mockRange(), this, rhs)
 
-fun greaterEqual(lhs: Expression, rhs: Expression) = lhs geq rhs
-
 infix fun Expression.geq(rhs: Expression) = OperatorBinary.GreaterEqual(mockRange(), this, rhs)
-
-fun greaterThan(lhs: Expression, rhs: Expression) = lhs gt rhs
 
 infix fun Expression.gt(rhs: Expression) = OperatorBinary.Greater(mockRange(), this, rhs)
 
-fun lnot(expr: Expression) = OperatorUnary.Negation(mockRange(), expr)
-
 infix fun Expression.addeq(rhs: Expression) = OperatorBinary.AdditionAssignment(mockRange(), this, rhs)
+
+fun minus(expression: Expression) = OperatorUnary.Minus(mockRange(), expression)
+
+fun lnot(expr: Expression) = OperatorUnary.Negation(mockRange(), expr)
 
 fun lit(int: Int) = Literal.IntLiteral(mockRange(), int)
