+++ src/test/kotlin/cacophony/controlflow/generation/SideEffectSeparationTest.kt
@@ -28,8 +28,8 @@ import cacophony.controlflow.generation.TestOperators.Companion.subNode
 import cacophony.controlflow.integer
 import cacophony.controlflow.rax
 import cacophony.controlflow.writeRegister
+import cacophony.functionDeclaration
 import cacophony.lit
-import cacophony.semantic.functionDeclaration
 import cacophony.variableDeclaration
 import cacophony.variableUse
 import cacophony.variableWrite
