+++ src/test/kotlin/cacophony/semantic/FunctionAnalysisTest.kt
@@ -1,5 +1,6 @@
 package cacophony.semantic
 
+import cacophony.*
 import cacophony.semantic.syntaxtree.Empty
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
@@ -175,7 +176,8 @@ class FunctionAnalysisTest {
                         AnalyzedFunction(
                             funF,
                             ParentLink(
-                                program, true,
+                                program,
+                                true,
                             ),
                             setOf(AnalyzedVariable(varA, program, VariableUseType.READ_WRITE)),
                             mutableSetOf(),
@@ -209,7 +211,8 @@ class FunctionAnalysisTest {
                         AnalyzedFunction(
                             funF,
                             ParentLink(
-                                program, false,
+                                program,
+                                false,
                             ),
                             emptySet(),
                             mutableSetOf(),
@@ -409,7 +412,8 @@ class FunctionAnalysisTest {
                         AnalyzedFunction(
                             funFoo,
                             ParentLink(
-                                program, false,
+                                program,
+                                false,
                             ),
                             setOf(AnalyzedVariable(fooVarA, funFoo, VariableUseType.READ)),
                             mutableSetOf(),
