+++ src/test/kotlin/cacophony/semantic/VarUseTypeAnalysisTest.kt
@@ -1,8 +1,8 @@
 package cacophony.semantic
 
+import cacophony.*
 import cacophony.semantic.syntaxtree.Empty
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
 
 class VarUseTypeAnalysisTest {
